+++ HabboHotel/Users/Effects/AvatarEffectFactory.cs
@@ -1,4 +1,6 @@
-﻿namespace Plus.HabboHotel.Users.Effects;
+﻿using Dapper;
+
+namespace Plus.HabboHotel.Users.Effects;
 
 internal static class AvatarEffectFactory
 {
@@ -11,11 +13,10 @@ internal static class AvatarEffectFactory
     /// <returns></returns>
     public static AvatarEffect CreateNullable(Habbo habbo, int spriteId, double duration)
     {
-        using var dbClient = PlusEnvironment.DatabaseManager.GetQueryReactor();
-        dbClient.SetQuery("INSERT INTO `user_effects` (`user_id`,`effect_id`,`total_duration`,`is_activated`,`activated_stamp`,`quantity`) VALUES(@uid,@sid,@dur,'0',0,1)");
-        dbClient.AddParameter("uid", habbo.Id);
-        dbClient.AddParameter("sid", spriteId);
-        dbClient.AddParameter("dur", duration);
-        return new(Convert.ToInt32(dbClient.InsertQuery()), habbo.Id, spriteId, duration, false, 0, 1);
+        using var connection = PlusEnvironment.DatabaseManager.Connection();
+        var id = Convert.ToInt32(connection.ExecuteScalar<long>(
+            "INSERT INTO `user_effects` (`user_id`,`effect_id`,`total_duration`,`is_activated`,`activated_stamp`,`quantity`) VALUES(@uid,@sid,@dur,'0',0,1); SELECT LAST_INSERT_ID();",
+            new { uid = habbo.Id, sid = spriteId, dur = duration }));
+        return new(id, habbo.Id, spriteId, duration, false, 0, 1);
     }
 }